requirements.txt
alembic==0.9.4
amqp==2.2.2
aniso8601==1.2.1
asn1crypto==0.24.0
billiard==3.5.0.3
blinker==1.4
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==1.9
defusedxml==0.5.0
dominate==2.3.1
eventlet==0.22.1
Flask==0.12.2
Flask-Admin==1.0.8
Flask-Bootstrap==3.3.7.1
Flask-Celery-Helper==1.1.0
Flask-Celery-py3==0.2.4
Flask-Dashboard==0.0.6a0
Flask-Excel==0.0.7
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.0
Flask-Moment==0.5.1
Flask-OAuthlib==0.9.4
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.5
Flask-SocketIO==2.9.6
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gevent==1.2.2
gevent-websocket==0.10.1
greenlet==0.4.13
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jwt==0.5.3
kombu==4.1.0
lml==0.0.1
Mako==1.0.7
MarkupSafe==1.0
numpy==1.13.3
oauth2==1.9.0.post1
oauthlib==2.0.2
pandas==0.22.0
pika==0.11.2
pycparser==2.18
pyexcel==0.5.3
pyexcel-io==0.4.4
pyexcel-webio==0.1.2
PyMySQL==0.8.0
pyserial==3.4
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==2.0.4
python-socketio==1.9.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
scapy-python3==0.23
six==1.10.0
SQLAlchemy==1.1.13
texttable==0.9.1
urllib3==1.22
vine==1.1.4
visitor==0.1.3
Werkzeug==0.12.2
WTForms==2.1
xerial==1.1.11b1
xlrd==1.1.0
xlwt==1.3.0
yunpian-python-sdk==1.0.0
